Increase version number to 5.8
Tested compilation of the dkms module with the patch script before and
after change.

Before the change an error occurred while compiling, it seemed that a
very old version of the hpsa.c file was used and therfore not able to
compile in the new version.
After the patch the module compiles successfully and works on a 5.8.0-1
truenas scale kernel.